"refactor: BackupHistory findAll() - totalElements"

+++ src/main/java/com/sprint/mission/sb8hrbankteamquerity/dto/BackupHistory/CursorPageResponseBackupHistoryDto.java
@@ -7,7 +7,7 @@ public record CursorPageResponseBackupHistoryDto(
     String nextCursor,
     Long nextIdAfter,
     int size,
-    Integer totalElements,
+    Long totalElements,
     boolean hasNext
 ) {
 }
